commit2

+++ task10/Program.cs
@@ -1,7 +1,7 @@
 ﻿int a = new Random().Next(100, 999);
 Console.Write("a = ");
 Console.WriteLine(a);
-double SecondNumber = 0;
+int SecondNumber = 0;
 
 Console.Write("SecondNumber = ");
 Console.WriteLine(SecondNumber = (a % 100 - a % 10) / 10);
